yoast_seo/streamlit_app/requirements.txt
streamlit
pandas
nltk
openai
python-dotenv
